LayoutTests/platform/win/accessibility/multiple-select-element-role-expected.txt

This tests the accessibility roles of multiselect elements and their option elements.

PASS accessibilityController.focusedElement.role is "list"
PASS accessibilityController.focusedElement.childAtIndex(0).role is "list item"
